SRE-50 Expose metrics on a separate port
This commit adds a separate http server for serving the metrics, rather
than sharing it with the api. Exposing the entire api is a security
issue since that exposes the raw backups over http as well.

Also added an additional metric containing the timestamp of the last run
for alerting purposes.